@@ -3,41 +3,24 @@
 # V3.17 required for the FindCUDAToolkit package
 # V3.18 required for CMAKE_CUDA_ARCHITECTURES
 cmake_minimum_required(VERSION 3.18)
-find_program(CMAKE_C_COMPILER NAMES $ENV{CC} gcc PATHS ENV PATH NO_DEFAULT_PATH)
-find_program(CMAKE_CXX_COMPILER NAMES $ENV{CXX} g++ PATHS ENV PATH NO_DEFAULT_PATH)
+
+if (USE_HIP)
+    find_program(CMAKE_HIP_COMPILER hipcc REQUIRED)
+    set(CMAKE_C_COMPILER ${CMAKE_HIP_COMPILER})
+    set(CMAKE_CXX_COMPILER ${CMAKE_HIP_COMPILER})
+endif()
 
 ## Project settings
-project(astaroth C CXX CUDA)
+project(astaroth C CXX)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR})
 
-# Needs to be here before setting the compilation flags for some unknown reason,
-# otherwise MPI not found
-if (MPI_ENABLED)
-    find_package(MPI REQUIRED COMPONENTS C)
-    include_directories(${MPI_C_INCLUDE_DIRS})
-    add_definitions(-DAC_MPI_ENABLED=1)
-endif ()
-
 ## Project-wide compilation flags
-set(COMMON_FLAGS "-DOMPI_SKIP_MPICXX -Wall -Wextra -Werror -Wdouble-promotion -Wfloat-conversion -Wshadow") # -DOMPI_SKIP_MPICXX is to force OpenMPI to use the C interface
+# -Werror TODO readd
+set(COMMON_FLAGS "-DOMPI_SKIP_MPICXX -Wall -Wextra -Wdouble-promotion -Wfloat-conversion -Wshadow -Wno-unused-result") #-DOMPI_SKIP_MPICXX is to force OpenMPI to use the C interface
 set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} ${COMMON_FLAGS}")
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${COMMON_FLAGS}")
 set(CMAKE_C_STANDARD 11)
-set(CMAKE_CXX_STANDARD 14)
-
-## CUDA
-# GPU, compute capability
-# K40, 3.5
-# K80, 3.7
-# P100, 6.0
-# V100, 7.0
-if (NOT CUDA_ARCHITECTURES)
-    set(CMAKE_CUDA_ARCHITECTURES 60 70) # Default
-else ()
-    set(CMAKE_CUDA_ARCHITECTURES ${CUDA_ARCHITECTURES}) # User-specified
-endif()
-string (REPLACE " " "," CUDA_COMMON_FLAGS "${COMMON_FLAGS}")
-set(CMAKE_CUDA_FLAGS "${CMAKE_CUDA_FLAGS} --compiler-options=${CUDA_COMMON_FLAGS}")
+set(CMAKE_CXX_STANDARD 17)
 
 ## Build type
 if(NOT CMAKE_BUILD_TYPE)
