@@ -1,5 +1,5 @@
 /*
-    Copyright (C) 2014-2020, Johannes Pekkila, Miikka Vaisala.
+    Copyright (C) 2014-2021, Johannes Pekkila, Miikka Vaisala.
 
     This file is part of Astaroth.
 
@@ -17,94 +17,110 @@
     along with Astaroth.  If not, see <http://www.gnu.org/licenses/>.
 */
 #pragma once
-#include <cuda_runtime_api.h> // Vector types
-#include <float.h>            // FLT_EPSILON, etc
-#include <stdio.h>            // printf
-#include <stdlib.h>           // size_t
 
-// Library flags
-#define STENCIL_ORDER (6)
-#define NGHOST (STENCIL_ORDER / 2)
-
-// Built-in types and parameters
-#if AC_DOUBLE_PRECISION == 1
-typedef double AcReal;
-typedef double3 AcReal3;
-#define AC_REAL_MAX (DBL_MAX)
-#define AC_REAL_MIN (DBL_MIN)
-#define AC_REAL_EPSILON (DBL_EPSILON)
-
-#if AC_MPI_ENABLED
-#define AC_MPI_TYPE MPI_DOUBLE
-#endif
-
-#else
-typedef float AcReal;
-typedef float3 AcReal3;
-#define AC_REAL_MAX (FLT_MAX)
-#define AC_REAL_MIN (FLT_MIN)
-#define AC_REAL_EPSILON (FLT_EPSILON)
-
-#if AC_MPI_ENABLED
-#define AC_MPI_TYPE MPI_FLOAT
-#endif
-
-#endif
+#include "acc_runtime.h"
+#define NGHOST (STENCIL_ORDER / 2) // Astaroth 2.0 backwards compatibility
 
 typedef struct {
-    AcReal3 row[3];
-} AcMatrix;
+    AcReal* vertex_buffer[NUM_VTXBUF_HANDLES];
+    AcMeshInfo info;
+} AcMesh;
 
-typedef enum { AC_SUCCESS = 0, AC_FAILURE = 1 } AcResult;
+#define STREAM_0 (0)
+#define STREAM_1 (1)
+#define STREAM_2 (2)
+#define STREAM_3 (3)
+#define STREAM_4 (4)
+#define STREAM_5 (5)
+#define STREAM_6 (6)
+#define STREAM_7 (7)
+#define STREAM_8 (8)
+#define STREAM_9 (9)
+#define STREAM_10 (10)
+#define STREAM_11 (11)
+#define STREAM_12 (12)
+#define STREAM_13 (13)
+#define STREAM_14 (14)
+#define STREAM_15 (15)
+#define STREAM_16 (16)
+#define STREAM_17 (17)
+#define STREAM_18 (18)
+#define STREAM_19 (19)
+#define STREAM_20 (20)
+#define STREAM_21 (21)
+#define STREAM_22 (22)
+#define STREAM_23 (23)
+#define STREAM_24 (24)
+#define STREAM_25 (25)
+#define STREAM_26 (26)
+#define STREAM_27 (27)
+#define STREAM_28 (28)
+#define STREAM_29 (29)
+#define STREAM_30 (30)
+#define STREAM_31 (31)
+#define NUM_STREAMS (32)
+#define STREAM_DEFAULT (STREAM_0)
+#define STREAM_ALL (NUM_STREAMS)
+typedef int Stream;
 
-// Device
-typedef struct device_s* Device; // Opaque pointer to device_s. Analogous to dispatchable handles
-                                 // in Vulkan, f.ex. VkDevice
-#define AC_KERNEL_FUNC_NAME(ID) acKernel_##ID
-#include "user_defines.h" // Autogenerated defines from the DSL
+#define AC_FOR_RTYPES(FUNC)                                                                        \
+    FUNC(RTYPE_MAX)                                                                                \
+    FUNC(RTYPE_MIN)                                                                                \
+    FUNC(RTYPE_RMS)                                                                                \
+    FUNC(RTYPE_RMS_EXP)                                                                            \
+    FUNC(RTYPE_ALFVEN_MAX)                                                                         \
+    FUNC(RTYPE_ALFVEN_MIN)                                                                         \
+    FUNC(RTYPE_ALFVEN_RMS)                                                                         \
+    FUNC(RTYPE_SUM)
+
+#define AC_FOR_BCTYPES(FUNC)                                                                       \
+    FUNC(BOUNDCOND_PERIODIC)                                                                       \
+    FUNC(BOUNDCOND_SYMMETRIC)                                                                      \
+    FUNC(BOUNDCOND_ANTISYMMETRIC)                                                                  \
+    FUNC(BOUNDCOND_A2)                                                                             \
+    FUNC(BOUNDCOND_PRESCRIBED_DERIVATIVE)
+
+#ifdef AC_INTEGRATION_ENABLED
+
+#define AC_FOR_SPECIAL_MHD_BCTYPES(FUNC)                                                           \
+    FUNC(SPECIAL_MHD_BOUNDCOND_ENTROPY_CONSTANT_TEMPERATURE)                                       \
+    FUNC(SPECIAL_MHD_BOUNDCOND_ENTROPY_BLACKBODY_RADIATION)                                        \
+    FUNC(SPECIAL_MHD_BOUNDCOND_ENTROPY_PRESCRIBED_HEAT_FLUX)                                       \
+    FUNC(SPECIAL_MHD_BOUNDCOND_ENTROPY_PRESCRIBED_NORMAL_AND_TURBULENT_HEAT_FLUX)
+
+#endif
+
+#define AC_FOR_INIT_TYPES(FUNC)                                                                    \
+    FUNC(INIT_TYPE_RANDOM)                                                                         \
+    FUNC(INIT_TYPE_AA_RANDOM)                                                                         \
+    FUNC(INIT_TYPE_XWAVE)                                                                          \
+    FUNC(INIT_TYPE_GAUSSIAN_RADIAL_EXPL)                                                           \
+    FUNC(INIT_TYPE_ABC_FLOW)                                                                       \
+    FUNC(INIT_TYPE_SIMPLE_CORE)                                                                    \
+    FUNC(INIT_TYPE_KICKBALL)                                                                       \
+    FUNC(INIT_TYPE_VEDGE)                                                                          \
+    FUNC(INIT_TYPE_VEDGEX)                                                                         \
+    FUNC(INIT_TYPE_RAYLEIGH_TAYLOR)                                                                \
+    FUNC(INIT_TYPE_RAYLEIGH_BENARD)
 
 #define AC_GEN_ID(X) X,
-
 // Naming the associated number of the boundary condition types
 typedef enum {
     AC_FOR_BCTYPES(AC_GEN_ID) //
     NUM_BCTYPES,
 } AcBoundcond;
 
+#ifdef AC_INTEGRATION_ENABLED
+typedef enum {
+    AC_FOR_SPECIAL_MHD_BCTYPES(AC_GEN_ID) //
+    NUM_SPECIAL_MHD_BCTYPES,
+} AcSpecialMHDBoundcond;
+#endif
+
 typedef enum {
     AC_FOR_RTYPES(AC_GEN_ID) //
     NUM_RTYPES
 } ReductionType;
-
-typedef enum {
-    AC_FOR_USER_INT_PARAM_TYPES(AC_GEN_ID) //
-    NUM_INT_PARAMS
-} AcIntParam;
-
-typedef enum {
-    AC_FOR_USER_INT3_PARAM_TYPES(AC_GEN_ID) //
-    NUM_INT3_PARAMS
-} AcInt3Param;
-
-typedef enum {
-    AC_FOR_USER_REAL_PARAM_TYPES(AC_GEN_ID) //
-    NUM_REAL_PARAMS
-} AcRealParam;
-
-typedef enum {
-    AC_FOR_USER_REAL3_PARAM_TYPES(AC_GEN_ID) //
-    NUM_REAL3_PARAMS
-} AcReal3Param;
-
-typedef enum {
-    AC_FOR_SCALARARRAY_HANDLES(AC_GEN_ID) //
-    NUM_SCALARARRAY_HANDLES
-} ScalarArrayHandle;
-
-typedef enum {
-    AC_FOR_VTXBUF_HANDLES(AC_GEN_ID) //
-    NUM_VTXBUF_HANDLES
-} VertexBufferHandle;
 
 typedef enum {
     AC_FOR_INIT_TYPES(AC_GEN_ID) //
@@ -115,35 +131,20 @@ typedef enum {
 
 #define _UNUSED __attribute__((unused)) // Does not give a warning if unused
 #define AC_GEN_STR(X) #X,
-static const char* bctype_names[] _UNUSED     = {AC_FOR_BCTYPES(AC_GEN_STR) "-end-"};
-static const char* rtype_names[] _UNUSED      = {AC_FOR_RTYPES(AC_GEN_STR) "-end-"};
-static const char* intparam_names[] _UNUSED   = {AC_FOR_USER_INT_PARAM_TYPES(AC_GEN_STR) "-end-"};
-static const char* int3param_names[] _UNUSED  = {AC_FOR_USER_INT3_PARAM_TYPES(AC_GEN_STR) "-end-"};
-static const char* realparam_names[] _UNUSED  = {AC_FOR_USER_REAL_PARAM_TYPES(AC_GEN_STR) "-end-"};
-static const char* real3param_names[] _UNUSED = {AC_FOR_USER_REAL3_PARAM_TYPES(AC_GEN_STR) "-end-"};
-static const char* scalarray_names[] _UNUSED  = {AC_FOR_SCALARARRAY_HANDLES(AC_GEN_STR) "-end-"};
-static const char* vtxbuf_names[] _UNUSED     = {AC_FOR_VTXBUF_HANDLES(AC_GEN_STR) "-end-"};
+static const char* bctype_names[] _UNUSED       = {AC_FOR_BCTYPES(AC_GEN_STR) "-end-"};
+static const char* rtype_names[] _UNUSED        = {AC_FOR_RTYPES(AC_GEN_STR) "-end-"};
 static const char* initcondtype_names[] _UNUSED = {AC_FOR_INIT_TYPES(AC_GEN_STR) "-end-"};
+
+#ifdef AC_INTEGRATION_ENABLED
+static const char* special_bctype_names[] _UNUSED = {
+    AC_FOR_SPECIAL_MHD_BCTYPES(AC_GEN_STR) "-end-"};
+#endif
+
 #undef AC_GEN_STR
 #undef _UNUSED
 
-typedef struct {
-    int int_params[NUM_INT_PARAMS];
-    int3 int3_params[NUM_INT3_PARAMS];
-    AcReal real_params[NUM_REAL_PARAMS];
-    AcReal3 real3_params[NUM_REAL3_PARAMS];
-} AcMeshInfo;
-
-typedef struct {
-    AcReal* vertex_buffer[NUM_VTXBUF_HANDLES];
-    AcMeshInfo info;
-} AcMesh;
-
-// Node
-typedef struct node_s* Node; // Opaque pointer to node_s.
-
-// Grid
-// typedef struct grid_s* Grid; // Opaque pointer to grid_s
+typedef struct node_s* Node;
+typedef struct device_s* Device;
 
 typedef struct {
     int3 m;
@@ -295,12 +296,14 @@ acQueryReal3params(void)
 }
 
 /** Prints a list of Scalar array handles */
+/*
 static inline void
 acQueryScalarrays(void)
 {
     for (int i = 0; i < NUM_SCALARARRAY_HANDLES; ++i)
         printf("%s (%d)\n", scalarray_names[i], i);
 }
+*/
 
 /** Prints a list of vertex buffer handles */
 static inline void
@@ -431,6 +434,12 @@ AcResult acGridLoadVectorUniform(const Stream stream, const AcReal3Param param,
                                  const AcReal3 value);
 
 /** */
+AcResult acGridLoadIntUniform(const Stream stream, const AcIntParam param, const int value);
+
+/** */
+AcResult acGridLoadInt3Uniform(const Stream stream, const AcInt3Param param, const int3 value);
+
+/** */
 AcResult acGridLoadMesh(const Stream stream, const AcMesh host_mesh);
 
 /** */
@@ -469,6 +478,14 @@ AcResult acGridReduceVecScal(const Stream stream, const ReductionType rtype,
                              const VertexBufferHandle vtxbuf2, const VertexBufferHandle vtxbuf3,
                              AcReal* result);
 
+typedef enum {
+    ACCESS_READ,
+    ACCESS_WRITE,
+} AccessType;
+
+AcResult acGridAccessMeshOnDisk(const VertexBufferHandle vtxbuf, const char* path,
+                                const AccessType type);
+
 /*
  * =============================================================================
  * Task interface (part of the grid interface)
@@ -476,9 +493,12 @@ AcResult acGridReduceVecScal(const Stream stream, const ReductionType rtype,
  */
 
 /** */
-typedef enum AcTaskType { TASKTYPE_COMPUTE, TASKTYPE_HALOEXCHANGE, TASKTYPE_BOUNDCOND } AcTaskType;
-
-#define BIT(pos) (1U << (pos))
+typedef enum AcTaskType {
+    TASKTYPE_COMPUTE,
+    TASKTYPE_HALOEXCHANGE,
+    TASKTYPE_BOUNDCOND,
+    TASKTYPE_SPECIAL_MHD_BOUNDCOND
+} AcTaskType;
 
 typedef enum AcBoundary {
     BOUNDARY_X_TOP = 0x01,
@@ -503,27 +523,50 @@ typedef struct AcTaskDefinition {
     union {
         AcKernel kernel;
         AcBoundcond bound_cond;
+#ifdef AC_INTEGRATION_ENABLED
+        AcSpecialMHDBoundcond special_mhd_bound_cond;
+#endif
     };
     AcBoundary boundary;
-    VertexBufferHandle* vtxbuf_dependencies;
-    size_t num_vtxbufs;
+
+    Field* fields_in;
+    size_t num_fields_in;
+
+    Field* fields_out;
+    size_t num_fields_out;
+
+    AcRealParam* parameters;
+    size_t num_parameters;
 } AcTaskDefinition;
 
 /** TaskGraph is an opaque datatype containing information necessary to execute a set of
- * operation.*/
+ * operations.*/
 typedef struct AcTaskGraph AcTaskGraph;
 
 /** */
-AcTaskDefinition acCompute(const AcKernel kernel, VertexBufferHandle vtxbuf_dependencies[],
-                           const size_t num_vtxbufs);
+AcTaskDefinition acCompute(const AcKernel kernel, Field fields_in[], const size_t num_fields_in,
+                           Field fields_out[], const size_t num_fields_out);
 
 /** */
-AcTaskDefinition acHaloExchange(VertexBufferHandle vtxbuf_dependencies[], const size_t num_vtxbufs);
+AcTaskDefinition acHaloExchange(Field fields[], const size_t num_fields);
 
 /** */
 AcTaskDefinition acBoundaryCondition(const AcBoundary boundary, const AcBoundcond bound_cond,
-                                     VertexBufferHandle vtxbuf_dependencies[],
-                                     const size_t num_vtxbufs);
+                                     Field fields[], const size_t num_fields,
+                                     AcRealParam parameters[], const size_t num_parameters);
+
+#ifdef AC_INTEGRATION_ENABLED
+/** SpecialMHDBoundaryConditions are tied to some specific DSL implementation (At the moment, the
+   MHD implementation). They launch specially written CUDA kernels that implement the specific
+   boundary condition procedure They are a stop-gap temporary solution. The sensible solution is to
+   replace them with a task type that runs a boundary condition procedure written in the Astaroth
+   DSL.
+*/
+AcTaskDefinition acSpecialMHDBoundaryCondition(const AcBoundary boundary,
+                                               const AcSpecialMHDBoundcond bound_cond,
+                                               AcRealParam parameters[],
+                                               const size_t num_parameters);
+#endif
 
 /** */
 AcTaskGraph* acGridGetDefaultTaskGraph();
@@ -535,7 +578,29 @@ AcTaskGraph* acGridBuildTaskGraph(const AcTaskDefinition ops[], const size_t n_o
 AcResult acGridDestroyTaskGraph(AcTaskGraph* graph);
 
 /** */
-AcResult acGridExecuteTaskGraph(const AcTaskGraph* graph, const size_t n_iterations);
+AcResult acGridExecuteTaskGraph(AcTaskGraph* graph, const size_t n_iterations);
+
+/** */
+AcResult acGridLaunchKernel(const Stream stream, const Kernel kernel, const int3 start,
+                            const int3 end);
+
+/** */
+AcResult acGridLoadStencil(const Stream stream, const Stencil stencil,
+                           const AcReal data[STENCIL_DEPTH][STENCIL_HEIGHT][STENCIL_WIDTH]);
+
+/** */
+AcResult acGridStoreStencil(const Stream stream, const Stencil stencil,
+                            AcReal data[STENCIL_DEPTH][STENCIL_HEIGHT][STENCIL_WIDTH]);
+
+/** */
+AcResult
+acGridLoadStencils(const Stream stream,
+                   const AcReal data[NUM_STENCILS][STENCIL_DEPTH][STENCIL_HEIGHT][STENCIL_WIDTH]);
+
+/** */
+AcResult
+acGridStoreStencils(const Stream stream,
+                    AcReal data[NUM_STENCILS][STENCIL_DEPTH][STENCIL_HEIGHT][STENCIL_WIDTH]);
 
 #endif // AC_MPI_ENABLED
 
@@ -700,7 +765,7 @@ AcResult acDeviceDestroy(Device device);
 AcResult acDevicePrintInfo(const Device device);
 
 /** */
-AcResult acDeviceAutoOptimize(const Device device);
+// AcResult acDeviceAutoOptimize(const Device device);
 
 /** */
 AcResult acDeviceSynchronizeStream(const Device device, const Stream stream);
@@ -728,9 +793,27 @@ AcResult acDeviceLoadInt3Uniform(const Device device, const Stream stream, const
                                  const int3 value);
 
 /** */
+AcResult acDeviceStoreScalarUniform(const Device device, const Stream stream,
+                                    const AcRealParam param, AcReal* value);
+
+/** */
+AcResult acDeviceStoreVectorUniform(const Device device, const Stream stream,
+                                    const AcReal3Param param, AcReal3* value);
+
+/** */
+AcResult acDeviceStoreIntUniform(const Device device, const Stream stream, const AcIntParam param,
+                                 int* value);
+
+/** */
+AcResult acDeviceStoreInt3Uniform(const Device device, const Stream stream, const AcInt3Param param,
+                                  int3* value);
+
+/** */
+/*
 AcResult acDeviceLoadScalarArray(const Device device, const Stream stream,
                                  const ScalarArrayHandle handle, const size_t start,
                                  const AcReal* data, const size_t num);
+                                 */
 
 /** */
 AcResult acDeviceLoadMeshInfo(const Device device, const AcMeshInfo device_config);
@@ -831,6 +914,23 @@ AcResult acDeviceReduceVecScal(const Device device, const Stream stream_type,
 /** */
 AcResult acDeviceRunMPITest(void);
 
+/** */
+AcResult acDeviceLaunchKernel(const Device device, const Stream stream, const Kernel kernel,
+                              const int3 start, const int3 end);
+
+/** */
+AcResult acDeviceLoadStencil(const Device device, const Stream stream, const Stencil stencil,
+                             const AcReal data[STENCIL_DEPTH][STENCIL_HEIGHT][STENCIL_WIDTH]);
+
+/** */
+AcResult acDeviceStoreStencil(const Device device, const Stream stream, const Stencil stencil,
+                              AcReal data[STENCIL_DEPTH][STENCIL_HEIGHT][STENCIL_WIDTH]);
+
+/** */
+AcResult acDeviceVolumeCopy(const Device device, const Stream stream,                     //
+                            const AcReal* in, const int3 in_offset, const int3 in_volume, //
+                            AcReal* out, const int3 out_offset, const int3 out_volume);
+
 /*
  * =============================================================================
  * Helper functions
@@ -854,29 +954,65 @@ AcResult acHostMeshDestroy(AcMesh* mesh);
 
 #ifdef __cplusplus
 #if AC_MPI_ENABLED
-/** */
-template <size_t num_vtxbufs>
+/** Backwards compatible interface, input fields = output fields*/
+template <size_t num_fields>
 AcTaskDefinition
-acCompute(AcKernel kernel, VertexBufferHandle (&vtxbuf_dependencies)[num_vtxbufs])
+acCompute(AcKernel kernel, Field (&fields)[num_fields])
 {
-    return acCompute(kernel, vtxbuf_dependencies, num_vtxbufs);
+    return acCompute(kernel, fields, num_fields, fields, num_fields);
+}
+
+template <size_t num_fields_in, size_t num_fields_out>
+AcTaskDefinition
+acCompute(AcKernel kernel, Field (&fields_in)[num_fields_in], Field (&fields_out)[num_fields_out])
+{
+    return acCompute(kernel, fields_in, num_fields_in, fields_out, num_fields_out);
 }
 
 /** */
-template <size_t num_vtxbufs>
-AcTaskDefinition acHaloExchange(VertexBufferHandle (&vtxbuf_dependencies)[num_vtxbufs])
+template <size_t num_fields> AcTaskDefinition acHaloExchange(Field (&fields)[num_fields])
 {
-    return acHaloExchange(vtxbuf_dependencies, num_vtxbufs);
+    return acHaloExchange(fields, num_fields);
 }
 
 /** */
-template <size_t num_vtxbufs>
+template <size_t num_fields>
 AcTaskDefinition
 acBoundaryCondition(const AcBoundary boundary, const AcBoundcond bound_cond,
-                    VertexBufferHandle (&vtxbuf_dependencies)[num_vtxbufs])
+                    Field (&fields)[num_fields])
 {
-    return acBoundaryCondition(boundary, bound_cond, vtxbuf_dependencies, num_vtxbufs);
+    return acBoundaryCondition(boundary, bound_cond, fields, num_fields, nullptr, 0);
 }
+
+/** */
+template <size_t num_fields, size_t num_parameters>
+AcTaskDefinition
+acBoundaryCondition(const AcBoundary boundary, const AcBoundcond bound_cond,
+                    Field (&fields)[num_fields], AcRealParam (&parameters)[num_parameters])
+{
+    return acBoundaryCondition(boundary, bound_cond, fields, num_fields, parameters,
+                               num_parameters);
+}
+
+#if AC_INTEGRATION_ENABLED
+/** */
+template <size_t num_fields>
+AcTaskDefinition
+acSpecialMHDBoundaryCondition(const AcBoundary boundary, const AcSpecialMHDBoundcond bound_cond)
+{
+    return acSpecialMHDBoundaryCondition(boundary, bound_cond, nullptr, 0);
+}
+
+/** */
+template <size_t num_fields, size_t num_parameters>
+AcTaskDefinition
+acSpecialMHDBoundaryCondition(const AcBoundary boundary, const AcSpecialMHDBoundcond bound_cond,
+                              AcRealParam (&parameters)[num_parameters])
+{
+    return acSpecialMHDBoundaryCondition(boundary, bound_cond, parameters, num_parameters);
+}
+
+#endif
 
 /** */
 template <size_t n_ops>
