@@ -1,5 +1,5 @@
 /*
-    Copyright (C) 2014-2020, Johannes Pekkila, Miikka Vaisala.
+    Copyright (C) 2014-2021, Johannes Pekkila, Miikka Vaisala.
 
     This file is part of Astaroth.
 
@@ -30,6 +30,10 @@
 
 #include "host_memory.h"
 #include "model_boundconds.h"
+
+#define inv_dsx ((AcReal)1. / get(AC_dsx))
+#define inv_dsy ((AcReal)1. / get(AC_dsy))
+#define inv_dsz ((AcReal)1. / get(AC_dsz))
 
 // Standalone flags
 #define LDENSITY (1)
@@ -174,7 +178,7 @@ derx(const int i, const int j, const int k, const ModelScalar* arr)
     for (int offset = 0; offset < STENCIL_ORDER + 1; ++offset)
         pencil[offset] = arr[IDX(i + offset - STENCIL_ORDER / 2, j, k)];
 
-    return first_derivative(pencil, get(AC_inv_dsx));
+    return first_derivative(pencil, inv_dsx);
 }
 
 static inline ModelScalar
@@ -185,7 +189,7 @@ derxx(const int i, const int j, const int k, const ModelScalar* arr)
     for (int offset = 0; offset < STENCIL_ORDER + 1; ++offset)
         pencil[offset] = arr[IDX(i + offset - STENCIL_ORDER / 2, j, k)];
 
-    return second_derivative(pencil, get(AC_inv_dsx));
+    return second_derivative(pencil, inv_dsx);
 }
 
 static inline ModelScalar
@@ -203,7 +207,7 @@ derxy(const int i, const int j, const int k, const ModelScalar* arr)
         pencil_b[offset] = arr[IDX(i + offset - STENCIL_ORDER / 2, j + STENCIL_ORDER / 2 - offset,
                                    k)];
 
-    return cross_derivative(pencil_a, pencil_b, get(AC_inv_dsx), get(AC_inv_dsy));
+    return cross_derivative(pencil_a, pencil_b, inv_dsx, inv_dsy);
 }
 
 static inline ModelScalar
@@ -221,7 +225,7 @@ derxz(const int i, const int j, const int k, const ModelScalar* arr)
         pencil_b[offset] = arr[IDX(i + offset - STENCIL_ORDER / 2, j,
                                    k + STENCIL_ORDER / 2 - offset)];
 
-    return cross_derivative(pencil_a, pencil_b, get(AC_inv_dsx), get(AC_inv_dsz));
+    return cross_derivative(pencil_a, pencil_b, inv_dsx, inv_dsz);
 }
 
 static inline ModelScalar
@@ -232,7 +236,7 @@ dery(const int i, const int j, const int k, const ModelScalar* arr)
     for (int offset = 0; offset < STENCIL_ORDER + 1; ++offset)
         pencil[offset] = arr[IDX(i, j + offset - STENCIL_ORDER / 2, k)];
 
-    return first_derivative(pencil, get(AC_inv_dsy));
+    return first_derivative(pencil, inv_dsy);
 }
 
 static inline ModelScalar
@@ -243,7 +247,7 @@ deryy(const int i, const int j, const int k, const ModelScalar* arr)
     for (int offset = 0; offset < STENCIL_ORDER + 1; ++offset)
         pencil[offset] = arr[IDX(i, j + offset - STENCIL_ORDER / 2, k)];
 
-    return second_derivative(pencil, get(AC_inv_dsy));
+    return second_derivative(pencil, inv_dsy);
 }
 
 static inline ModelScalar
@@ -261,7 +265,7 @@ deryz(const int i, const int j, const int k, const ModelScalar* arr)
         pencil_b[offset] = arr[IDX(i, j + offset - STENCIL_ORDER / 2,
                                    k + STENCIL_ORDER / 2 - offset)];
 
-    return cross_derivative(pencil_a, pencil_b, get(AC_inv_dsy), get(AC_inv_dsz));
+    return cross_derivative(pencil_a, pencil_b, inv_dsy, inv_dsz);
 }
 
 static inline ModelScalar
@@ -272,7 +276,7 @@ derz(const int i, const int j, const int k, const ModelScalar* arr)
     for (int offset = 0; offset < STENCIL_ORDER + 1; ++offset)
         pencil[offset] = arr[IDX(i, j, k + offset - STENCIL_ORDER / 2)];
 
-    return first_derivative(pencil, get(AC_inv_dsz));
+    return first_derivative(pencil, inv_dsz);
 }
 
 static inline ModelScalar
@@ -283,14 +287,14 @@ derzz(const int i, const int j, const int k, const ModelScalar* arr)
     for (int offset = 0; offset < STENCIL_ORDER + 1; ++offset)
         pencil[offset] = arr[IDX(i, j, k + offset - STENCIL_ORDER / 2)];
 
-    return second_derivative(pencil, get(AC_inv_dsz));
+    return second_derivative(pencil, inv_dsz);
 }
 
 #if LUPWD
 static inline ModelScalar
 der6x_upwd(const int i, const int j, const int k, const ModelScalar* arr)
 {
-    ModelScalar inv_ds = get(AC_inv_dsx);
+    ModelScalar inv_ds = inv_dsx;
 
     return ModelScalar(1.0 / 60.0) * inv_ds *
            (-ModelScalar(20.0) * arr[IDX(i, j, k)] +
@@ -302,7 +306,7 @@ der6x_upwd(const int i, const int j, const int k, const ModelScalar* arr)
 static inline ModelScalar
 der6y_upwd(const int i, const int j, const int k, const ModelScalar* arr)
 {
-    ModelScalar inv_ds = get(AC_inv_dsy);
+    ModelScalar inv_ds = inv_dsy;
 
     return ModelScalar(1.0 / 60.0) * inv_ds *
            (-ModelScalar(20.0) * arr[IDX(i, j, k)] +
@@ -314,7 +318,7 @@ der6y_upwd(const int i, const int j, const int k, const ModelScalar* arr)
 static inline ModelScalar
 der6z_upwd(const int i, const int j, const int k, const ModelScalar* arr)
 {
-    ModelScalar inv_ds = get(AC_inv_dsz);
+    ModelScalar inv_ds = inv_dsz;
 
     return ModelScalar(1.0 / 60.0) * inv_ds *
            (-ModelScalar(20.0) * arr[IDX(i, j, k)] +
