@@ -1,5 +1,5 @@
 /*
-    Copyright (C) 2014-2020, Johannes Pekkila, Miikka Vaisala.
+    Copyright (C) 2014-2021, Johannes Pekkila, Miikka Vaisala.
 
     This file is part of Astaroth.
 
@@ -144,10 +144,13 @@ update_config(AcMeshInfo* config)
     config->int_params[AC_nz_max] = config->int_params[AC_nz] + STENCIL_ORDER / 2;
 
     // Spacing
+    /*
+    // %JP: AC_inv_ds[xyz] now calculated inside the mhd kernel
     config->real_params[AC_inv_dsx] = AcReal(1.) / config->real_params[AC_dsx];
     config->real_params[AC_inv_dsy] = AcReal(1.) / config->real_params[AC_dsy];
     config->real_params[AC_inv_dsz] = AcReal(1.) / config->real_params[AC_dsz];
-    config->real_params[AC_dsmin]   = min(config->real_params[AC_dsx],
+    */
+    config->real_params[AC_dsmin] = min(config->real_params[AC_dsx],
                                         min(config->real_params[AC_dsy],
                                             config->real_params[AC_dsz]));
 
