@@ -1,3 +1,21 @@
+/*
+    Copyright (C) 2014-2021, Johannes Pekkila, Miikka Vaisala.
+
+    This file is part of Astaroth.
+
+    Astaroth is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    Astaroth is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with Astaroth.  If not, see <http://www.gnu.org/licenses/>.
+*/
 #pragma once
 #include "astaroth.h"
 
@@ -7,15 +25,6 @@
 
 #define MPI_GPUDIRECT_DISABLED (0)
 #endif // AC_MPI_ENABLED
-
-typedef AcReal AcRealPacked;
-
-typedef struct {
-    AcReal* in[NUM_VTXBUF_HANDLES];
-    AcReal* out[NUM_VTXBUF_HANDLES];
-
-    AcReal* profiles[NUM_SCALARARRAY_HANDLES];
-} VertexBufferArray;
 
 struct device_s {
     int id;
@@ -30,18 +39,21 @@ struct device_s {
     AcReal* reduce_result;
 };
 
+typedef AcReal AcRealPacked;
+
 typedef struct {
+    Kernel kernel;
     cudaStream_t stream;
     int step_number;
     int3 start;
     int3 end;
 } KernelParameters;
 
-typedef AcResult (*ComputeKernel)(const KernelParameters, VertexBufferArray);
-
 #ifdef __cplusplus
 extern "C" {
 #endif
+
+// Generic boundconds
 
 /** */
 AcResult acKernelPeriodicBoundconds(const cudaStream_t stream, const int3 start, const int3 end,
@@ -52,6 +64,46 @@ AcResult acKernelSymmetricBoundconds(const cudaStream_t stream, const int3 regio
                                      const int3 normal, const int3 dims, AcReal* vtxbuf);
 
 /** */
+AcResult acKernelAntiSymmetricBoundconds(const cudaStream_t stream, const int3 region_id,
+                                         const int3 normal, const int3 dims, AcReal* vtxbuf);
+
+/** */
+AcResult acKernelA2Boundconds(const cudaStream_t stream, const int3 region_id, const int3 normal,
+                              const int3 dims, AcReal* vtxbuf);
+
+/** */
+AcResult acKernelPrescribedDerivativeBoundconds(const cudaStream_t stream, const int3 region_id,
+                                                const int3 normal, const int3 dims, AcReal* vtxbuf,
+                                                AcRealParam der_val_param);
+
+// Entropy boundconds
+
+#ifdef AC_INTEGRATION_ENABLED
+/** */
+AcResult acKernelEntropyConstantTemperatureBoundconds(const cudaStream_t stream,
+                                                      const int3 region_id, const int3 normal,
+                                                      const int3 dims, VertexBufferArray vba);
+
+/** */
+AcResult acKernelEntropyBlackbodyRadiationKramerConductivityBoundconds(const cudaStream_t stream,
+                                                                       const int3 region_id,
+                                                                       const int3 normal,
+                                                                       const int3 dims,
+                                                                       VertexBufferArray vba);
+
+/** */
+AcResult acKernelEntropyPrescribedHeatFluxBoundconds(const cudaStream_t stream,
+                                                     const int3 region_id, const int3 normal,
+                                                     const int3 dims, VertexBufferArray vba,
+                                                     AcRealParam F_param);
+
+AcResult acKernelEntropyPrescribedNormalAndTurbulentHeatFluxBoundconds(
+    const cudaStream_t stream, const int3 region_id, const int3 normal, const int3 dims,
+    VertexBufferArray vba, AcRealParam hcond_param, AcRealParam F_param);
+
+#endif
+
+/** */
 AcResult acKernelGeneralBoundconds(const cudaStream_t stream, const int3 start, const int3 end,
                                    AcReal* vtxbuf, const VertexBufferHandle vtxbuf_handle,
                                    const AcMeshInfo config, const int3 bindex);
@@ -60,14 +112,19 @@ AcResult acKernelGeneralBoundconds(const cudaStream_t stream, const int3 start, 
 AcResult acKernelDummy(void);
 
 /** */
-AcResult acKernelAutoOptimizeIntegration(const int3 start, const int3 end, VertexBufferArray vba);
+AcResult acKernelFlush(AcReal* arr, const size_t n);
+
+/** */
+// AcResult acKernelAutoOptimizeIntegration(const int3 start, const int3 end,
+// VertexBufferArray vba);
 
 /** */
 AcResult acKernelPackData(const cudaStream_t stream, const VertexBufferArray vba,
                           const int3 vba_start, const int3 dims, AcRealPacked* packed);
 
 /** */
-AcResult acKernelIntegrateSubstep(const KernelParameters params, VertexBufferArray vba);
+// AcResult acKernelIntegrateSubstep(const KernelParameters params,
+// VertexBufferArray vba);
 
 /** */
 AcResult acKernelPartialPackData(const cudaStream_t stream, const VertexBufferArray vba,
@@ -99,10 +156,13 @@ AcReal acKernelReduceVecScal(const cudaStream_t stream, const ReductionType rtyp
                              const AcReal* vtxbuf2, const AcReal* vtxbuf3, AcReal* scratchpad,
                              AcReal* reduce_result);
 
-#define GEN_KERNEL_FUNC_DECL(ID)                                                                   \
-    AcResult AC_KERNEL_FUNC_NAME(ID)(const KernelParameters params, VertexBufferArray vba);
+/** */
+AcResult acKernelVolumeCopy(const cudaStream_t stream,                                    //
+                            const AcReal* in, const int3 in_offset, const int3 in_volume, //
+                            AcReal* out, const int3 out_offset, const int3 out_volume);
 
-#include "user_kernel_decl.h"
+// Astaroth 2.0 backwards compatibility.
+AcResult acKernel(const KernelParameters params, VertexBufferArray vba);
 
 #ifdef __cplusplus
 } // extern "C"
